Add MetaExtractor class
Add new MetaExtractor class to handle reading and writing MetaFields
to and from an external database.

This class uses similar subset of methods to the MetaEditor class,
with a focus on returning a custom collection of MetaField objects.

These collections are used for serializing/deserializing tags to
an external database.