Add IMetaStoreReader class
Add new IMetaStoreReader interface.

This interface is similar to the IMetaFieldVisitor interface, with the
exception that the methods return a MetaField object instead of
void.

This is necessary as the MetaStoreReader class collects each MetaField
object before it sends them off to be serialized and inserted into an
external database.